Article Recommendation System processed with NLP
I preprocessed the krapivin 2009 data, which was processed with NLP enhanced with a functional interface, with spacy. Afterwards, I returned the results of the words searched according to the vectors of interest in the database according to FastText and SCIBERT models. There is also an editable profile page and article search screen.